remove comsole color dep and write to consoler source
